requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
httpx==0.25.2
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.24.3
motor==3.3.2
pymongo==4.6.0